Modification de la classe action (ajout des Float sans cost)

+++ actions_class.py
@@ -3,7 +3,7 @@ class Action:
 
     def __init__(self, action):
         self.name = action[0]
-        self.cost = int(action[1])
+        self.cost = float(action[1])
         self.profit = round(self.cost * float(action[2][:-1]) / 100, 2)
 
     def __str__(self) -> str:
